requirements.txt
python-dotenv
openai
requests
flask 